initial prach implementation. Splitting of components for RU (FFTs) and L1 (detection). Update of UE to include all formats (for 30 kHz SCS case on initial BWP)

+++ nfapi/open-nFAPI/nfapi/public_inc/nfapi_nr_interface.h
@@ -528,7 +528,7 @@ typedef struct{
   uint16_t   srs_monitoring_periodicity;
   uint16_t   slot_monitoring_periodicity;
   uint16_t   slot_monitoring_offset;
-  uint16_t  monitoring_symbols_in_slot;
+  uint32_t  monitoring_symbols_in_slot;
   uint16_t  number_of_candidates[NFAPI_NR_MAX_NB_CCE_AGGREGATION_LEVELS];
 } nfapi_nr_search_space_t;
